+++ extensions/publisher-command-center/src/models/Processes.js
@@ -23,12 +23,14 @@ const Processes = {
         this._fetch = null;
         throw err;
       });
+
+    return this._fetch;
   },
 
-  destroy: function (content_id, process_id) {
+  destroy: function (contentId, processId) {
     return m.request({
       method: "DELETE",
-      url: `api/contents/${content_id}/processes/${process_id}`,
+      url: `api/contents/${contentId}/processes/${processId}`,
     });
   },